+++ livekit-datatrack/CHANGELOG.md
@@ -1,4 +1,17 @@
 # Changelog
+## 0.1.10 (2026-07-09)
+
+### Fixes
+
+- Handle data track SID reassignment
+- introduce LiveKitAPI construct, added smoke tests - #1220 (@davidzhao)
+
+## 0.1.9 (2026-06-23)
+
+### Fixes
+
+- Upgrade protocol to v1.48.0
+
 ## 0.1.8 (2026-05-29)
 
 ### Fixes